Update draft-ietf-oauth-transaction-tokens.md
Relaxed the language around the definition of the value passed in the `subject_token`. The value for `subject_token` can range from a simple string to a complex token as identified by the value of `subject_token_type`. This edit still makes the `subject_token` and `subject_token_type` required parameters.

This addresses issue #53