Removal of `purp` claim
Based on conversations and issue comments in issues #178 and #194, this PR renames the `purp` claim in the Transaction Token to `scope`. Clarifying text has been added to make it clear that the value of the `scope` claim in the Transaction Token does NOT need to be the same value requested or present in a provided access_token. 

Please review for clarity of the added text.

I also removed the IANA registry for 'purp'.